+++ modules/admin/AdminModule.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace app\modules\admin;
+
+use yii\base\Module;
+
+class AdminModule extends Module
+{
+    /**
+     * {@inheritdoc}
+     */
+    public $defaultRoute = 'admin/statistic';
+}