+++ modules/admin/controllers/BaseController.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace app\modules\admin\controllers;
+
+use yii\web\Controller;
+use yii\filters\AccessControl;
+use yii\filters\VerbFilter;
+
+class BaseController extends Controller
+{
+    /**
+     * {@inheritdoc}
+     */
+    public $layout = 'main';
+
+    /**
+     * {@inheritdoc}
+     */
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::class,
+                'rules' => [
+                    [
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
+            'verbs' => [
+                'class' => VerbFilter::class,
+                'actions' => [
+                    'delete' => ['post'],
+                ],
+            ],
+        ];
+    }
+}